Savestate related fixes for MusicModuleMIDI/JVCMSXMIDI
MusicModuleMIDI + JVCMSXMIDI:
All MSXDevice subclasses need to call
   ar.template serializeBase<MSXDevice>(*this);
in their serialize() implementation.

JVCMSXMIDI:
Detail: serialize version = 1 is the default, no need to explicitly mention
that.

MC6850:
Even though this is a new class, still keep the serialize version = 3 to keep
it possible to load (very) old savestates that have 'MusicModuleMIDI' version 1
or 2.

MusicModuleMIDI:
Add backwards compatibility serialization code. This was different from our
usual backwards compatibility code: in this case the same hardware description
results in different instantiated C++ classes. But luckily the "polymorphic
type" (a string) that's stored in the savestate is different for both cases.
And then it was possible to register both names for this (new) class. And
instead of starting at version 1 for this new class we continue incrementing
from the old version number (old was 3, so now we're at 4).

serialize_meta.cc:
More robust against invalid savestate files (print an error instead of
trigger an assertion). Shouldn't happen anymore with the above fixes for
backwards compatibility, but cleaner anyway.